Clean up copy-pasted code
